requirements.txt
blinker==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
flask==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
Flask-Cors==4.0.1
gunicorn==22.0.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
